Do display margin notes on narrow screens.
In contrast to the upstream tufte-css, I believe we should display
margin notes by default, even on narrow viewports. Nobody on a phone
display is going to be clicking the tiny superscript note numbers. In my
opinion, it is closer to the original intent to display the note by
default so the reader doesn't have to go hunting for it.

Accomplish this by checking the toggle checkboxes by default. If you
want to change this behavior, you can manually write the HTML and remove
the checked attribute.