Move plugins to standard ruby gem lib paths.
It turns out that Jekyll themes do not load plugins from _plugins when a
gem-based theme is used.

So as a result, we need to turn this gem into an actual Ruby gem that
can be require'd in order to load the plugin library code in sites that
depend on it.

Symlink all the files into _plugins so that they can still be used in
the demo site.